Correção das Funções de Regressão Linear

+++ src/main/kotlin/regressao/Registro.kt
@@ -1,5 +1,5 @@
 package regressao
 
-data class Registro(val xPetro: Float,
-                    val xDolar: Float,
-                    val yPreco: Float)
+data class Registro(val xPetro: Double,
+                    val xDolar: Double,
+                    val yPreco: Double)
